Enhance `nats_kv` processor (#2495)
* Add `keys_filter` field to nats_kv processor

This is a slightly-breaking change to the nats_kv processor which
is still beta. It alters the behaviour of the `keys` operation
which now returns an array of keys instead of a batch of messages,
as discussed with @codegangsta.

I'm a bit concerned that the history operation still returns a
batch of messages, which might confuse people. Happy to modify
that one as well to return an array of objects if you'd like. In
the current implementation, some of the fields of each history
item are emitted as message metadata, which is a bit odd anyway,
since they're not identical across the batch...

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Log error in debug mode if closing the key watcher fails

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Switch the `nats_kv` components to the new jetstream API

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Refactor the `nats_kv` `keys` operation

Use the `key` field for filtering keys instead of adding an extra
`keys_filter` field.

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Adjust the `nats_kv` `history` operation

Return an array of objects instead of a batch of messages.

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Add `timeout` field to the `nats_kv` processor

Signed-off-by: Mihai Todor <todormihai@gmail.com>

* Update the changelog

Signed-off-by: Mihai Todor <todormihai@gmail.com>

---------

Signed-off-by: Mihai Todor <todormihai@gmail.com>
Co-authored-by: Ashley Jeffs <ash@jeffail.uk>